Update Mapper Object Api

+++ ProjectTLCNShopCore/Areas/Admin/Controllers/UsersController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using ProjectTLCNShopCore.EF;
 using Microsoft.EntityFrameworkCore;
 using ProjectTLCNShopCore.Areas.Admin.Dao;
+using Microsoft.AspNetCore.Http;
 
 // For more information on enabling MVC for empty projects, visit http://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -24,12 +25,20 @@ namespace ProjectTLCNShopCore.Areas.Admin.Controllers
 		// GET: Users
 		public async Task<IActionResult> Index()
 		{
+			if (HttpContext.Session.GetString("UserID") == null && HttpContext.Session.GetString("Email") == null)
+			{
+				return RedirectToAction("Index", "Login");
+			}
 			return View(await _context.Users.ToListAsync());
 		}
 
 		// GET: Users/Details/5
 		public async Task<IActionResult> Details(int? id)
 		{
+			if (HttpContext.Session.GetString("UserID") == null && HttpContext.Session.GetString("Email") == null)
+			{
+				return RedirectToAction("Index", "Login");
+			}
 			if (id == null)
 			{
 				return NotFound();
@@ -47,6 +56,10 @@ namespace ProjectTLCNShopCore.Areas.Admin.Controllers
 		// GET: Users/Create
 		public IActionResult Create()
 		{
+			if (HttpContext.Session.GetString("UserID") == null && HttpContext.Session.GetString("Email") == null)
+			{
+				return RedirectToAction("Index", "Login");
+			}
 			return View();
 		}
 
@@ -57,6 +70,10 @@ namespace ProjectTLCNShopCore.Areas.Admin.Controllers
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> Create([Bind("UserId,AvatarPicture,Birthday,CoverPicture,Email,FirstName,IsActive,LastName,Looked,Lpoint,Password,PhoneNumber,ReferUserId,Sex,Token")] Users users)
 		{
+			if (HttpContext.Session.GetString("UserID") == null && HttpContext.Session.GetString("Email") == null)
+			{
+				return RedirectToAction("Index", "Login");
+			}
 			if (ModelState.IsValid)
 			{
 				_context.Add(users);
@@ -69,6 +86,10 @@ namespace ProjectTLCNShopCore.Areas.Admin.Controllers
 		// GET: Users/Edit/5
 		public async Task<IActionResult> Edit(int? id)
 		{
+			if (HttpContext.Session.GetString("UserID") == null && HttpContext.Session.GetString("Email") == null)
+			{
+				return RedirectToAction("Index", "Login");
+			}
 			if (id == null)
 			{
 				return NotFound();
@@ -89,6 +110,10 @@ namespace ProjectTLCNShopCore.Areas.Admin.Controllers
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> Edit(int id, [Bind("UserId,AvatarPicture,Birthday,CoverPicture,Email,FirstName,IsActive,LastName,Looked,Lpoint,Password,PhoneNumber,ReferUserId,Sex,Token")] Users users)
 		{
+			if (HttpContext.Session.GetString("UserID") == null && HttpContext.Session.GetString("Email") == null)
+			{
+				return RedirectToAction("Index", "Login");
+			}
 			if (id != users.UserId)
 			{
 				return NotFound();
@@ -120,6 +145,10 @@ namespace ProjectTLCNShopCore.Areas.Admin.Controllers
 		// GET: Users/Delete/5
 		public async Task<IActionResult> Delete(int? id)
 		{
+			if (HttpContext.Session.GetString("UserID") == null && HttpContext.Session.GetString("Email") == null)
+			{
+				return RedirectToAction("Index", "Login");
+			}
 			if (id == null)
 			{
 				return NotFound();
@@ -139,6 +168,10 @@ namespace ProjectTLCNShopCore.Areas.Admin.Controllers
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> DeleteConfirmed(int id)
 		{
+			if (HttpContext.Session.GetString("UserID") == null && HttpContext.Session.GetString("Email") == null)
+			{
+				return RedirectToAction("Index", "Login");
+			}
 			var users = await _context.Users.SingleOrDefaultAsync(m => m.UserId == id);
 			_context.Users.Remove(users);
 			await _context.SaveChangesAsync();
